requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0
joblib>=1.4.0
scikit-learn>=1.5.0
xgboost>=2.1.0
pandas>=2.2.0
numpy>=1.26.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
